Doan 17

+++ src/Controller/UserProjectsController.php
@@ -44,19 +44,37 @@ class UserProjectsController extends AppController
     }
 
     public function add($project_id) {
-        if ($this->request->is('post')) {
-            $userProject = $this->UserProjects->find()->where([
-                'user_id' => $_POST['user_id'],
-                'project_id' => $project_id,
-            ])->first();
-            if ($userProject == array() && $_POST['status'] == 'true') {
-                $this->UserProjects->addNew($_POST['user_id'], $project_id);
-            } if ($userProject != array() && $_POST['status'] == 'false') {
-                $a = $this->UserProjects->get($userProject['id']);
-                $this->UserProjects->delete($a);die();
+        if ($this->current_user['position_id'] == 1 || $this->current_user['position_id'] == 2 || $this->current_user['position_id'] == 5 ) {
+            if ($this->request->is('post')) {
+                $userProject = $this->UserProjects->find()->where([
+                    'user_id' => $_POST['user_id'],
+                    'project_id' => $project_id,
+                ])->first();
+                if ($userProject == array() && $_POST['status'] == 'true') {
+                    $this->UserProjects->addNew($_POST['user_id'], $project_id);
+                } if ($userProject != array() && $_POST['status'] == 'false') {
+                    $a = $this->UserProjects->get($userProject['id']);
+                    $this->UserProjects->delete($a);die();
+                }
+                return($this->redirect('/UserProjects/index/'.$project_id));
             }
-            return($this->redirect('/UserProjects/index/'.$project_id));
+        } else {
+            return($this->redirect('/error/error404/'));
         }
+    }
+
+    // lấy danh sách dự án đang tham gia của user hiện tại
+    public function getListProjectsJoin() {
+        $projects = $this->UserProjects->find()->where(['user_id' => $this->current_user['id']])->contain(['Projects' => function($q){
+            return $q->contain(['companies']);
+        }]);
+        $position_current_user = $this->current_user['position_id'];
+        $this->set(compact('projects', 'position_current_user'));
+        // return $this->find()->where(['ProjectTeams.team_id' => $team_id])->contain(['projects' => function($q){
+        //     return $q->contain('Companies');
+        // }])->innerJoinWith('Teams', function($q) use($user_id) {
+        //     return $q->where(['Teams.leader' => $user_id]);
+        // });
     }
 }
 ?>
